tao_matcher/requirements.txt
# Django + extensions
Django==4.1.2
django-axes==5.39.0
django-bitfield==2.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-richtextfield==1.6.1
django-storages==1.13.1
django-tables2==2.4.1

crispy-bootstrap5==0.7

# Others
boto3==1.24.89
gunicorn==20.1.0
imbalanced-learn==0.9.0
matplotlib==3.6.1
mysqlclient==2.1.1
newspaper3k==0.2.8
openai==0.23.1
openpyxl==3.0.10
pdfkit==1.0.0
python-dotenv==0.20.0
random-user-agent==1.0.1
ratelimit==2.2.1
redis==4.3.4
requests==2.28.1
scikit-learn==1.0.2
seaborn==0.12.0
selenium==4.5.0
XlsxWriter==3.0.3

# git+https://${GFM_UTILS}@github.com/globalfundmedia/gfm-utils.git
